[refactor] jyanken_go and jyanken module
