+++ go.mod
@@ -1,0 +1,3 @@
+module jyanken_go
+
+go 1.17
